+++ schematics/src/ng-new/schema.d.ts
@@ -1,0 +1,17 @@
+export interface Schema {
+    directory?: string;
+    name: string;
+    skipInstall?: boolean;
+    linkCli?: boolean;
+    skipGit?: boolean;
+    commit?: { name: string, email: string, message?: string } | boolean;
+    newProjectRoot?: string;
+    version?: string;
+    inlineStyle?: boolean;
+    inlineTemplate?: boolean;
+    viewEncapsulation?: ('Emulated' | 'Native' | 'None');
+    routing?: boolean;
+    prefix?: string;
+    style?: string;
+    skipTests?: boolean;
+}