+++ schematics/src/utils/index.ts
@@ -1,4 +1,4 @@
-import { SchematicsException, Tree } from '@angular-devkit/schematics';
+import { Rule, SchematicsException, Tree } from '@angular-devkit/schematics';
 
 export function getSource(tree: Tree, filePath: string): string {
     const text = tree.read(filePath);
@@ -8,4 +8,29 @@ export function getSource(tree: Tree, filePath: string): string {
     }
 
     return text.toString('utf-8');
+}
+
+export function createGitIgnore(dirName: string): Rule {
+    return (tree => {
+        tree.create(`./${dirName}/.gitignore`, `
+/node_modules/
+/dist/
+/lib/
+/yarn.lock
+*.log
+.idea
+.serverless
+*.iml
+*.js.map
+*.d.ts
+.DS_Store
+dll
+.awcache
+/src/styles/main.css
+/firebug-lite
+firebug-lite.tar.tgz
+serverless.yml
+/coverage`);
+        return tree;
+    });
 }