+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 1.1.2
-* Window-ref on server side - [fix #151](https://github.com/maciejtreder/angular-universal-pwa/issues/151)
+* Window-ref on server side [fix #151](https://github.com/maciejtreder/angular-universal-pwa/issues/151)
 
 ### 2.0.0-beta.0
 * update to Angular 6 RC.3 (code only) - [0ba1d07](https://github.com/maciejtreder/angular-universal-pwa/commit/0ba1d07e792b4043a2bb8b045cb1dc3bd4bbb9dc])
@@ -7,3 +7,8 @@
 ### 2.0.0-beta.1
 * update code to Angular 6 RC.4 (code only)
 * build fix - [#209](https://github.com/maciejtreder/angular-universal-pwa/issues/209)
+
+### 2.0.0-beta.2
+* firebug is no longer downloaded with schematics [fix #211](https://github.com/maciejtreder/angular-universal-pwa/issues/211)
+* firebug is downloaded only if is not present on the path [fix #204](https://github.com/maciejtreder/angular-universal-pwa/issues/204)
+* Angular 6 CLI support (ng-new collection) [fix #205](https://github.com/maciejtreder/angular-universal-pwa/issues/205)